requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
APScheduler==3.10.1
astor==0.8.1
astroid==2.15.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
chardet @ file:///tmp/build/80754af9/chardet_1605303192865/work
charset-normalizer==3.1.0
click==8.1.3
conda==4.14.0
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1605484832764/work
contourpy==1.1.0
cryptography==40.0.2
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dicttoxml==1.7.16
dill==0.3.6
dnspython==2.3.0
docstring-to-markdown==0.12
executing==1.2.0
fake-useragent==1.1.3
flake8==6.0.0
Flask==2.3.2
Flask-Cors==3.0.10
fonttools==4.40.0
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
google==3.0.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-tools==1.54.0
httpie==3.2.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jwcrypto==1.4.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
limits==3.4.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
neovim==0.3.1
numpy==1.25.0
openai
packaging==23.1
pandas==1.5.3
pandasai==0.6.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.22.3
ptpython==3.0.23
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycosat==0.6.3
pycparser==2.21
pycryptodomex==3.9.4
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==1.7.1
pylint==2.17.2
pymongo==4.3.3
pynvim==0.4.3
pyOpenSSL==23.1.1
pyparsing==3.1.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
pytoolconfig==1.2.5
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.5.5
requests==2.30.0
requests-toolbelt==0.10.1
rich==13.3.4
rope==1.7.0
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1605527293619/work
selenium==2.53.6
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
stack-data==0.6.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
traitlets==5.9.0
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
Unidecode==1.3.6
urllib3==1.26.15
uWSGI @ file:///croot/uwsgi_1678466245296/work
vcrpy==4.3.1
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.4
whatthepatch==1.0.4
wrapt==1.15.0
xmltodict==0.13.0
yapf==0.32.0
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
